tests to a separate mod; benchmarks; rust update
